pyproject.toml
[project]
name = "git-mcp"
version = "0.1.0"
description = "MCP server for Git operations"
requires-python = ">=3.9"
dependencies = [
    "mcp>=0.9.0",
    "GitPython>=3.1.0",
    "pydantic>=2.0.0",
]

[project.scripts]
git-mcp = "git_mcp:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"